TMS_HW2526/pyproject.toml
[tool.poetry]
name = "default"
version = "0.1.0"
description = ""
authors = ["I admin <dirks.anton96@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
sqlalchemy = "^2.0.23"
sqlalchemy-utils = "^0.41.1"
mysql = "^0.0.3"
mysql-connector-python = "^8.2.0"
django-ckeditor = "^6.7.0"
pillow = "^10.2.0"
django-cleanup = "^8.0.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
